break out precision scan script and add more plots
