+++ .gitignore
@@ -1,0 +1,7 @@
+*csv
+*png
+test*
+*~
+*pycache*
+data
+log*
